backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pypdf2
pymupdf
langchain
langchain-google-genai
google-generativeai
GoogleNews
python-dotenv
textblob
newspaper
yfinance